wrote row to postgres
